backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
sentence-transformers==3.1.1
faiss-cpu==1.13.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3